fixed type error in readReg32Bit(); fixes #4 (thanks ScruffR)

+++ VL53L0X.cpp
@@ -343,7 +343,7 @@ uint16_t VL53L0X::readReg16Bit(uint8_t reg)
 // Read a 32-bit register
 uint32_t VL53L0X::readReg32Bit(uint8_t reg)
 {
-  uint16_t value;
+  uint32_t value;
 
   Wire.beginTransmission(address);
   Wire.write(reg);
